GH-1745 - Improve RetryTopicConfigurer flexibility
